Re-worked model classes for HW1
I separated the two and three card match games into separate model classes. All I need to do in the ThreeCardMatchingGame subclass of CardMatchingGame is override flipCardAtIndex: with the 3 card logic.

This makes the game much more expandable as I can just plug in a new model class for each type of game and leave the existing game types alone.